Housekeeping to meet R CMD CHECK
There are several outstanding notes and warnings, but all except one are
related to the format of the example data provided in `data`. R packages
prefer this to be data in R binary formats that is documented in the
package. This can either be moved to `inst/extdata` as preferred by the
R package guidelines, or kept in `data` which is more user friendly