Regression test outputs test results
You can now run `R/tests/benchmark.R` to test whether the
`extractFeatures` and `varsFromTimeSeries` functions still produce the
same values as they did prior to the large dataframe oriented refactor.
Previously this script was designed to be run interactively but now it
prints its results and returns an appropriate status code (0 if all
tests pass, 1 otherwise) so it can be used in an automated fashion.