minBox uses the first pair of points in a tie
minBox calculates the pairwise distance between all coordinates to find
the minimum enclosing box. Previously the way in which the coordinate
pair was determined in case of a tie in the pairwise distances was
undetermined. The function has been rewritten so that it is now the
first pair with the maximum distance that is chosen.