Add tolerance when taking floor in cooccurrence matrix
The rescaling can result in values that mathematically should be equal
to whole integers, say 4, but when stored as floating point numbers it
can be 3.99999999999996, which when floored will round down to 3, when
it should be 4.
This commit fixes this behaviour by adding a small tolerance to the
pixel value before flooring.