fix to trajectory area denominator (#22)
The denominator is the number of frames that a cell appeared in to
normalise the area covered to the duration of the cell being tracked.
Previously this didn't take interpolated frames into account. I.e. if a
cell appeared in frames 1, 2, 3, (missing), 6, 7, 8, the denominator
would be 6, rather than the 8 frames that the cell was actually moving
for. This commit fixes this