Cluster labels are correctly assigned and plot displays
Previously the first instance assigned to each cluster wasn't assigned a
label in the output object. So if k=3, 3 samples wouldn't be classified.
This was due to an indexing problem.

Also fixes an issue where the plot wasn't correctly being displayed.
This is resolved by calling the `print.ggplot2` method to force it to
display.